update size of device and volume
git-svn-id: svn+ssh://140.110.240.196/partclone_svn@209 93b75dfe-da2d-4241-a08d-2141769abc04

+++ unstable/src/ntfsclone-ng.c
@@ -51,7 +51,7 @@ ntfs_volume *ntfs;
 static void fs_open(char* device){
    
     int	    err;
-    int	    device_size, volume_size;
+    unsigned long long int	    device_size, volume_size;
 
     ntfs = ntfs_mount(device, NTFS_MNT_RDONLY);
     if (!ntfs) {
